Add scripts and templates for managing dev-install hosting
The utilities and templates provide preliminary dev-install automation
for hosting providers. In order to use this feature as currently
implemented, the Koha code must be checked out to /home/hosting/kohaclone
and the regular installer has to be run to create a dev install in the
directory /home/hosting/koha-dev/shared/ and then disabled.