Add beginnings of a simple cataloging interface
This initial commit only includes support for editing existing records.
Subjects are not currently handled. Titles are parsed based on ISBD
punctuation. The author heading is authorized with an AJAX-y connection
to the JSON authority look-up script.  The three notes fields can be
duplicated and removed, but at present there is no check to make sure
that there is one of each field left, so it is possible to remove all
the notes boxes. This interface requires Template::Toolkit. It does not
use the Template::Pro template interface used by the rest of Koha. There
is also no authentication support at the moment. Anyone can access the
simple cataloging interface if they know the URL. This will change. The
following keyword links are needed for the simple cataloging interface
to work (examples for MARC21 are given):
* author_authid -> 100$9
* author_name -> 100$a
* author_authid -> 110$9
* author_name -> 110$a
* titleproper -> 245$a
* subtitle -> 245$b
* responsibility -> 245$c
* city -> 260$a
* publisher -> 260$b
* date -> 260$c
* pagination -> 300$a
* size -> 300$c
* notes -> 500$a
* provenance -> 561$a
* binding -> 563$a