Bug 7284 follow up: Cataloging module relinking
With the first patch for bug 7284, the automatic authority linking will actually
work properly in the cataloging module. As Owen pointed out while testing the
patch, though, longtime users of Koha will not be expecting that. In keeping
with the principles of least surprise and maximum configurability, this patch
makes it possible to disable authority relinking in the cataloging module only
(i.e. leaving it enabled for future runs of link_bibs_to_authorities.pl).

This patch addds the following syspref:
* CatalogModuleRelink - when turned on, the automatic linker will relink
  headings when a record is saved in the cataloging module when LinkerRelink
  is turned on, even if the headings were manually linked to a different
  authority by the cataloger. When turned off (the default), the automatic
  linker will not relink any headings that have already been linked when a
  record is saved.

Note that though the default behavior matches the current behavior of Koha,
it does not match the intended behavior. Libraries that want the intended
behavior rather than the current behavior will need to adjust the
CatalogModuleRelink syspref.

Be sure to run the atomicupdate file used by this patch if you are on a dev
system: installer/data/mysql/atomicupdate/bug_7284_authority_linking_pt2

To test this patch:
1.  Run installer/data/mysql/atomicupdate/bug_7284_authority_linking_pt2
2.  Set BiblioAddsAuthorities to "on."
3.  Default setting of CatalogModuleRelink is "off." Leave it like that.
4.  Create a record and link an authority record to an authorized field using
    the authority plugin.
5.  Save the record. Ensure that the heading is linked to the appropriate
    authority.
6.  Open the record. Change the heading manually to something else, leaving
    the link. Save the record.
7.  Ensure that the heading remains linked to that same authority.
8.  Change CatalogModuleRelink to "on."
9.  Open the record. Use the authority plugin to link that heading to the
    same authority record you did earlier.
10. Save the record. Ensure that the heading is linked to the appropriate
    authority.
11. Open the record. Change the heading manually to something else, leaving
    the link. Save the record.
12. Ensure that the heading is no longer linked to the old authority record.