Bug 9135: updatedatabase.pl fails when DEBUG is set
When the DEBUG environment variable is set to a true value, all DBI
errors are treated as fatal. Unfortunately, duplicate keys are
incredibly common in updatedatabase.pl, since sysprefs are often added
in testing, then re-added once the patches have been pushed to master
(or added when on a maintenance branch, and then re-added when the
system is upgraded to either a newer maintenance branch or master).
This patch disables fatal errors for the updatedatabase.pl script.