Bug 9658: Incorrect breadcrumb on report dictionary
Prior to this patch, the breadcrumb on the reports dictionary looks
like this: Home > Reports > Guided reports wizard Dictionary (notice
the lack of punctuation between "wizard" and "Dictionary").
This patch makes it consistent with the rest of the system so it looks
like this: Home > Reports > Guided reports wizard > Dictionary

I left dictionary bolded, since that seems to have been intentional,
even if I can't find a good reason for it.

To test:
1) View reports dictionary page before and after patch is applied,
   observing difference.